requirements.txt
# Dépendances principales
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
psutil==5.9.6

# Dépendances pour les tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
pytest-html==4.1.1
coverage==7.3.2
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1
